webMAN MOD 1.43.01
- Support  4.76 CEX (thanks Sony, for this biggest fw update...)
- Control fan system is better than before (more explanation below by
aldotools)
- The game list is now sort by CD# (e.g FF PS1)
- Local IP with adapter is now displayed (config and popup)
- Offsets improvement for fix game compatibility
- Updated Italian & Russian lang files

Let me try to explain how it works:
- webMAN's fan control tries to keep cool CPU and RSX.
- If one of them reach the target/max temperature (e.g. 66 C), the fan
speed will be increased gradually (+2) until the temperature start to
decrease.
- If the temperature is 1 degree or more higher than the target/max
temperature, then the fan speed is increased even more faster.
- When the temperature start to decrease, the fan speed is decreased
gradually (-1). This is done that way, because we want to keep the
temperature as cool as possible and avoid the "yo-yo" effect.

deank added additional heuristics to manage some special conditions, but
in simple terms this is the algorithm.

If you want to test it, I have uploaded a test build that decrease the
fan speed faster under the following conditions:
- If the temperature is 3 degrees below the target/max temperature, it
decreases the fan speed -1. if the fan speed is higher than 66%, it
decreases -2
- If the temperature is 5 degrees below the target/max temperature, it
decreases the fan speed -1. if the fan speed is higher than 50%, it
decreases -2

webMAN MOD already had something like these conditions, but with
different values (5 and 8 degrees, and 75% and 60%).